Changed: module names, manifest package - imageview - textview
