+++ admiral-build-src/src/main/kotlin/Publishing.kt
@@ -13,9 +13,9 @@ object Publishing {
         val uiKitCommon = "admiral-uikit-common"
         val uiKit = "admiral-uikit"
         val uiKitCompose = "admiral-uikit-compose"
-        val imageview = "admiral-imageview"
-        val textview = "admiral-textview"
-        val notification = "admiral-notification"
+        val imageview = "admiral-uikit-imageview"
+        val textview = "admiral-uikit-textview"
+        val notification = "admiral-uikit-notification"
         val uikitCore = "admiral-uikit-core"
     }
