--- admiral-imageview/build.gradle.kts
@@ -1,19 +1,0 @@
-plugins {
-    id("com.android.library")
-    id("org.jetbrains.kotlin.android")
-}
-
-android {
-    applyConfig(
-        isCoreLibraryDesugaringEnabled = true,
-        isComposeEnabled = false,
-        isViewBindingEnabled = true
-    )
-}
-
-dependencies {
-    api(project(Modules.Admiral.UiKit.core))
-
-    implementation(Libs.AndroidX.coreKtx)
-    implementation(Libs.AndroidX.appcompat)
-}
