--- admiral-notification/src/main/res/layout/admiral_notification_action.xml
@@ -1,90 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/admiral_notification_action"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:minHeight="52dp"
-    android:paddingStart="@dimen/module_x4"
-    android:paddingTop="@dimen/module_x3"
-    android:paddingEnd="@dimen/module_x4"
-    android:paddingBottom="@dimen/module_x3"
-    android:background="@drawable/admiral_bg_rectangle_8dp">
-
-    <FrameLayout
-        android:id="@+id/timerContainer"
-        android:layout_width="28dp"
-        android:layout_height="28dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <ProgressBar
-            android:id="@+id/timerProgressBar"
-            style="?android:attr/progressBarStyleHorizontal"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:indeterminate="false"
-            android:max="100"
-            android:progress="75"
-            android:progressDrawable="@drawable/admiral_progress_bar"
-            android:scaleX="-1"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <com.admiral.uikit.textview.TextView
-            android:id="@+id/timerText"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:gravity="center"
-            app:admiralTextStyle="subhead3"
-            tools:text="5" />
-
-    </FrameLayout>
-
-    <androidx.constraintlayout.widget.Barrier
-        android:id="@+id/barrier"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:barrierDirection="start"
-        app:constraint_referenced_ids="cancelImage,cancelText" />
-
-    <com.admiral.uikit.links.Link
-        android:id="@+id/cancelText"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="8dp"
-        app:admiralText="Отмена"
-        android:visibility="visible"
-        app:admiralTextStyle="body2"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <com.admiral.imageview.ImageView
-        android:id="@+id/cancelImage"
-        android:layout_width="28dp"
-        android:layout_height="28dp"
-        android:visibility="gone"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:visibility="visible" />
-
-    <com.admiral.uikit.textview.TextView
-        android:id="@+id/message"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/module_x3"
-        android:layout_marginEnd="@dimen/module_x2"
-        app:admiralTextStyle="body2"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="@id/barrier"
-        app:layout_constraintStart_toEndOf="@id/timerContainer"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_goneMarginEnd="0dp"
-        app:layout_goneMarginStart="0dp"
-        tools:text="Сообщение будет удалено" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
