+++ admiral-platform/build.gradle.kts
@@ -12,6 +12,9 @@ dependencies {
         api(project(Modules.Admiral.UiKit.uikit))
         api(project(Modules.Admiral.UiKit.common))
         api(project(Modules.Admiral.UiKit.compose))
+        api(project(Modules.Admiral.UiKit.Components.notification))
+        api(project(Modules.Admiral.UiKit.Components.imageview))
+        api(project(Modules.Admiral.UiKit.Components.textview))
     }
 }
