+++ admiral-uikit-imageview/proguard-rules.pro
