+++ admiral-uikit-imageview/src/androidTest/java/com/admiral/imageview/ExampleInstrumentedTest.kt
