+++ admiral-uikit-notification/src/main/res/layout/admiral_notification_static.xml
