+++ settings.gradle.kts
@@ -32,8 +32,8 @@ include(
     ":admiral-uikit-common",
     ":demo"
 )
-include(":admiral-imageview")
+include(":admiral-uikit-imageview")
 include(":admiral-uikit-core")
 include(":demo-separated-components")
-include(":admiral-textview")
-include(":admiral-notification")
+include(":admiral-uikit-textview")
+include(":admiral-uikit-notification")
